mnist-triton/requirements.txt
torch
torchvision
tritonclient[all]
attrdict
fastapi
starlette
hypercorn
tritonclient
python-multipart
numpy
Pillow
git+https://github.com/NVIDIA-AI-IOT/torch2trt@master